+++ src/cadqueryhelper/shape/__init__.py
@@ -17,3 +17,4 @@ from .cross import cross
 from .diamond import diamond
 from .chevron import chevron
 from .lightning import lightning
+from .cornerJoin import corner_join